+++ NEWS.md
@@ -3,7 +3,8 @@
  * Add functions to manage CNAG projects for `cellranger multi`.
  * Make `llrs_cellranger_clean()` more useful (keeping metrics).
  * Create folders just once in `llrs_cnag_symlinks()`
- * Set up environment variables in `llrs_profile()`
+ * Set up environment variables in `llrs_profile()`.
+ * New function `llrs_check_pkg_version()` to compare github version with local version.
  
 # rutils 0.0.1